requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
django-jet==1.0.8
djangorestframework==3.11.0
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
# git+https://github.com/mago960806/django-rest-framework-docs.git
django-oauth-toolkit
django-cors-middleware
sendgrid
stripe
django-map-widgets==0.2.0
djangorestframework-gis